automatically backed up by learn

+++ db/migrate/20180111145807_user.rb
@@ -3,7 +3,7 @@ class User < ActiveRecord::Migration[5.1]
     create_table :users do |c|
       c.string :username
       c.string :password
-      c.float :balance
+      c.decimal :balance
     end
   end
 end
